# docs/api-server/openapi.yaml
openapi: "3.1.0"
info:
  title: AgentClash API
  version: 1.0.0
  description: >
    B2B AI agent evaluation platform. Pits agent deployments against challenge
    packs, with replay inspection, scoring, comparison, and release gating.
  license:
    name: MIT
  contact:
    url: https://github.com/Atharva-Kanherkar/agentclash
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /artifacts/{artifactID}/content:
    get:
      operationId: getArtifactContent
      tags: [Artifacts]
      summary: Download artifact content
      description: >
        Streams artifact binary content. Requires a valid HMAC-signed URL
        obtained from the authenticated download endpoint. A 30-second grace
        period is allowed past the expiry timestamp.
      security:
        - artifactSignature: []
      parameters:
        - $ref: "#/components/parameters/ArtifactID"
        - name: expires
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: Unix timestamp when the signed URL expires
        - name: signature
          in: query
          required: true
          schema:
            type: string
          description: HMAC-SHA256 signature
      responses:
        "200":
          description: Artifact binary content
          headers:
            Content-Type:
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
            Content-Security-Policy:
              schema:
                type: string
                example: "sandbox; default-src 'none'"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ---------------------------------------------------------------------------
  # Hosted integration
  # ---------------------------------------------------------------------------
  /v1/integrations/hosted-runs/{runID}/events:
    post:
      operationId: ingestHostedRunEvent
      tags: [HostedRuns]
      summary: Ingest hosted run event
      description: >
        Webhook-style endpoint for externally hosted agents to report execution
        events. Authenticated via an HMAC-signed callback token (not JWT).
      security:
        - hostedRunToken: []
      parameters:
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostedRunEvent"
      responses:
        "202":
          description: Event accepted
          content:
            application/json:
              schema:
                type: object
                required: [accepted]
                properties:
                  accepted:
                    type: boolean
                    example: true
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"

  # ---------------------------------------------------------------------------
  # Session
  # ---------------------------------------------------------------------------
  /v1/auth/session:
    get:
      operationId: getSession
      tags: [Session]
      summary: Get current session
      description: Returns the authenticated user's identity and workspace memberships.
      security:
        - bearerJWT: []
      responses:
        "200":
          description: Current user session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  # ---------------------------------------------------------------------------
  # Workspace
  # ---------------------------------------------------------------------------
  /v1/workspaces/{workspaceID}/auth-check:
    get:
      operationId: checkWorkspaceAccess
      tags: [Workspace]
      summary: Verify workspace access
      description: Returns success if the caller has access to the specified workspace.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Workspace access confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceAccessCheckResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

  # ---------------------------------------------------------------------------
  # User Profile
  # ---------------------------------------------------------------------------
  /v1/users/me:
    get:
      operationId: getUserProfile
      tags: [UserProfile]
      summary: Get current user profile
      description: >
        Returns the authenticated user's profile with full org/workspace
        hierarchy including roles at each level.
      security:
        - bearerJWT: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Onboarding
  # ---------------------------------------------------------------------------
  /v1/onboarding:
    post:
      operationId: onboard
      tags: [Onboarding]
      summary: First-time onboarding
      description: >
        Creates a new organization and workspace for a first-time user.
        Fails if the user already has an organization.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardRequest"
      responses:
        "201":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Organizations
  # ---------------------------------------------------------------------------
  /v1/organizations:
    get:
      operationId: listOrganizations
      tags: [Organizations]
      summary: List organizations
      description: Lists organizations the caller belongs to.
      security:
        - bearerJWT: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListOrganizationsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createOrganization
      tags: [Organizations]
      summary: Create an organization
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/organizations/{organizationID}:
    get:
      operationId: getOrganization
      tags: [Organizations]
      summary: Get an organization
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/OrganizationID"
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    patch:
      operationId: updateOrganization
      tags: [Organizations]
      summary: Update an organization
      description: Requires org_admin role. Setting status to archived cascades to all workspaces.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/OrganizationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
...
  /v1/organization-memberships/{membershipID}:
    patch:
      operationId: updateOrgMembership
      tags: [OrgMemberships]
      summary: Update an organization membership
      description: >
        Org admins can change role or status. Invited users can accept
        (transition invited to active). Invite expires after 7 days.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/MembershipID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgMembershipRequest"
      responses:
        "200":
          description: Membership updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMembershipResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "410":
          $ref: "#/components/responses/GoneError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"

  # ---------------------------------------------------------------------------
  # Workspace Details
  # ---------------------------------------------------------------------------
  /v1/workspaces/{workspaceID}/details:
    get:
      operationId: getWorkspace
      tags: [Workspaces]
      summary: Get workspace details
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
    patch:
      operationId: updateWorkspace
      tags: [Workspaces]
      summary: Update workspace details
      description: Requires workspace_admin or org_admin. Setting status to archived cascades.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"

  # ---------------------------------------------------------------------------
  # Workspace Memberships
  # ---------------------------------------------------------------------------
  /v1/workspaces/{workspaceID}/memberships:
    get:
      operationId: listWorkspaceMemberships
      tags: [WorkspaceMemberships]
      summary: List workspace memberships
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWorkspaceMembershipsResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
    post:
      operationId: inviteWorkspaceMember
      tags: [WorkspaceMemberships]
      summary: Invite a member to a workspace
      description: Requires workspace_admin or org_admin. User must have active org membership first.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteWorkspaceMemberRequest"
      responses:
        "201":
          description: Member invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMembershipDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
...
  /v1/workspaces/{workspaceID}/secrets:
    get:
      operationId: listWorkspaceSecrets
      tags: [WorkspaceSecrets]
      summary: List workspace secret keys
      description: Returns secret metadata only — values are never exposed.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: List of secret keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWorkspaceSecretsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/runtime-profiles/{profileID}:
    get:
      operationId: getRuntimeProfile
      tags: [RuntimeProfiles]
      summary: Get a runtime profile
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/ProfileID"
      responses:
        "200":
          description: Runtime profile details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeProfileResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
...
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT Bearer token. In development mode, use custom headers instead:
        X-Agentclash-User-Id, X-Agentclash-WorkOS-User-Id,
        X-Agentclash-User-Email, X-Agentclash-User-Display-Name,
        X-Agentclash-Workspace-Memberships.
    hostedRunToken:
      type: http
      scheme: bearer
      description: >
        HMAC-SHA256 callback token issued when a hosted run is started.
        Not a JWT — the token encodes run_id and run_agent_id claims.
    artifactSignature:
      type: apiKey
      in: query
      name: signature
      description: >
        HMAC-SHA256 signature for artifact content URLs. Must be paired with
        an "expires" query parameter (unix timestamp).

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    ArtifactID:
      name: artifactID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...